+++ apps/formula-app/src/pages/Table.svelte
@@ -1,153 +1,134 @@
 <script>
-  import { get, writable } from 'svelte/store';
+  import { get } from 'svelte/store';
   import { beaker, formula } from '../../../../dist/packages/svelte/formula';
-
 
   const { form, formValues, updateForm } = formula();
 
+  // This creates a contact group - you can now bind `contacts.group` to the subgroup
+  const customers = beaker();
 
-  const contacts = beaker();
-  const contactValues = contacts.stores.formValues;
-  console.log(contacts);
+  export let productData = {
+    productName: '',
+  };
 
-  const formData = {
-    foo: 'Foo',
-    bar: true
-  }
+  // Set the store with any existing data
+  export let contactData = [{
+    firstName: '',
+    lastName: '',
+    email: '',
+    subscriptionLevel: '',
+    signups: [],
+  }];
+  const contactStore = customers.formValues;
+  const contactsValid = customers.validity;
+  contactStore.set(contactData);
 
-  const rows = writable([{
-    firstName: 'Tane',
-    lastName: 'Piper',
-    email: 'formula@svelte.codes',
-    checkboxMulti: ['B', 'C'],
-    checkbox: true,
-    radio: 'A',
-    selectOne: '2',
-    selectMulti: ['1', '3']
-  }, {
-    firstName: 'Foo',
-    lastName: 'McBarr',
-    email: 'formula@svelte.codes',
-    checkboxMulti: ['A', 'D'],
-    checkbox: true,
-    radio: 'B',
-    selectOne: '3',
-    selectMulti: ['1', '2']
-  }, {
-    firstName: 'Third',
-    lastName: 'Man',
-    email: 'formula1@svelte.codes',
-    checkboxMulti: ['A', 'C'],
-    checkbox: true,
-    radio: 'C',
-    selectOne: '1',
-    selectMulti: ['2', '3']
-  }]);
+  $: console.log($contactsValid);
 
-  $: console.log('formValues', $formValues);
-
-  function addRow() {
-    rows.update(state => [...state, {
+  // Add a row to the store
+  function addCustomer() {
+    customers.formValues.update(state => [...state, {
       firstName: '',
       lastName: '',
       email: '',
+      subscriptionLevel: '',
+      signups: [],
     }]);
   }
 
-  function deleteRow(index) {
-    rows.update(state => {
+  // Remove a row from the store
+  function deleteCustomer(index) {
+    customers.formValues.update(state => {
       state.splice(index, 1);
       return state;
     });
   }
 
-  function onSubmit() {
-    const formData = get(formValues);
-    const groupData = get(contacts.formValues);
-    console.log(formData, groupData);
+  function submit() {
+    const mainForm = get(formValues);
+    const contacts = get(contactStore);
+    //Do something with the data here
+    console.log(mainForm, contacts);
   }
-
-  const group = contacts.stores.formValues;
-  $: console.log('contactValues', $contactValues);
 </script>
 
-<form use:form on:submit|preventDefault={onSubmit}>
-  <label for='foo'>
-    Foo
-  </label>
-  <input type='text' id='foo' name='foo' bind:value={formData.foo}/>
+<form use:form on:submit={submit}>
+  <label for='productName'>ProductName</label>
+  <input type='text' id='productName' name='productName' bind:value={productData.productName} />
 
-  <input type='checkbox' id='bar' name='bar' bind:checked={formData.bar}/>
+  <button type='submit'>Submit Form</button>
+  <button on:click|preventDefault={addCustomer}>Add Customer</button>
 
-  <button on:click|preventDefault={addRow}>Add Row</button>
-  <button type='submit'>Submit</button>
   <table>
     <thead>
     <tr>
-      <th>First Name</th>
-      <th>Last Name</th>
-      <th>Email</th>
-      <th>Checkbox</th>
-      <th>Checkbox Multi</th>
-      <th>Radio</th>
-      <th>Select</th>
-      <th>Select Multi</th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th>Subscription Level</th>
+      <th>Subscriptions</th>
       <th>&nbsp;</th>
     </tr>
     </thead>
-    <tbody use:contacts.group id='contacts'>
-    {#each $rows as row, i}
+    <tbody use:customers.group>
+    {#each $contactStore as row, i}
       <tr>
         <td>
+          <label for='firstName-{i}'>First Name</label>
           <input type='text' id='firstName-{i}' name='firstName' required bind:value={row.firstName} />
         </td>
         <td>
-          <input type='text' id='lastName-{i}' name='lastName' bind:value={row.lastName}/>
+          <label for='lastName-{i}'>Last Name</label>
+          <input type='text' id='lastName-{i}' name='lastName' bind:value={row.lastName} required />
         </td>
         <td>
-          <input type='text' id='email-{i}' name='email' bind:value={row.email}/>
+          <label for='email-{i}'>Email Name</label>
+          <input type='email' id='email-{i}' name='email' bind:value={row.email} required />
         </td>
         <td>
-          <input type='checkbox' id='checkbox-{i}' name='checkbox' bind:checked={row.checkbox} />
-        </td>
-        <td>
-          <input type='checkbox' id='checkbox-multi-{i}-1' name='checkboxMulti' value='A'
-                 bind:group={row.checkboxMulti} />
-          <input type='checkbox' id='checkbox-multi-{i}-2' name='checkboxMulti' value='B'
-                 bind:group={row.checkboxMulti} />
-          <input type='checkbox' id='checkbox-multi-{i}-3' name='checkboxMulti' value='C'
-                 bind:group={row.checkboxMulti} />
-          <input type='checkbox' id='checkbox-multi-{i}-4' name='checkboxMulti' value='D'
-                 bind:group={row.checkboxMulti} />
-        </td>
-        <td>
-          <input type='radio' id='radio-{i}-1' name='radio-{i}' data-beaker-key='radio' value='A' bind:group={row.radio} />
-          <input type='radio' id='radio-{i}-2' name='radio-{i}' data-beaker-key='radio' value='B' bind:group={row.radio} />
-          <input type='radio' id='radio-{i}-3' name='radio-{i}' data-beaker-key='radio' value='C' bind:group={row.radio} />
-          <input type='radio' id='radio-{i}-4' name='radio-{i}' data-beaker-key='radio' value='D' bind:group={row.radio} />
-        </td>
-        <td>
-          <select name='selectOne' bind:value={row.selectOne}>
-            <option value='1'>1</option>
-            <option value='2'>2</option>
-            <option value='3'>3</option>
+          <!-- In multi-group forms, radio groups require a unique name in the DOM - her you can provide 'data-beaker-key' to specify the data key -->
+          <label for='subscriptionLevel-{i}-1'>None
+            <input type='radio' id='subscriptionLevel-{i}-1'
+                   name='subscriptionLevel-{i}'
+                   data-beaker-key='subscriptionLevel' value='none'
+                   bind:group={row.subscriptionLevel} />
+          </label>
 
-          </select>
-        </td>
-        <td>
-          <select name='selectMulti' multiple bind:value={row.selectMulti}>>
-            <option value='1'>1</option>
-            <option value='2'>2</option>
-            <option value='3'>3</option>
+          <label for='subscriptionLevel-{i}-1'>Partial
+            <input type='radio' id='subscriptionLevel-{i}-2'
+                   name='subscriptionLevel-{i}'
+                   data-beaker-key='subscriptionLevel' value='partial'
+                   bind:group={row.subscriptionLevel} />
+          </label>
 
-          </select>
+          <label for='subscriptionLevel-{i}-1'>Full
+            <input type='radio' id='subscriptionLevel-{i}-3'
+                   name='subscriptionLevel-{i}'
+                   data-beaker-key='subscriptionLevel' value='full'
+                   bind:group={row.subscriptionLevel} />
+          </label>
+
         </td>
+        <label for='signups-{i}-1'>Daily <input type='checkbox' id='signups-{i}-1' name='signups' value='daily'
+                                                bind:group={row.signups} /></label>
+
+        <label for='signups-{i}-2'>Weekly
+          <input type='checkbox' id='signups-{i}-2' name='signups' value='weekly'
+                 bind:group={row.signups} />
+        </label>
+        <label for='signups-{i}-3'>News
+          <input type='checkbox' id='signups-{i}-3' name='signups' value='news'
+                 bind:group={row.signups} />
+        </label>
+        <label for='signups-{i}-4'>Product
+          <input type='checkbox' id='signups-{i}-4' name='signups' value='product'
+                 bind:group={row.signups} />
+        </label>
         <td>
-          <button on:click|preventDefault={() => deleteRow(i)}>X</button>
+          <button on:click|preventDefault={() => deleteCustomer(i)}>X</button>
         </td>
       </tr>
     {/each}
     </tbody>
   </table>
 </form>
-